if disconnect for too long, no more fetchs

+++ ft_transcendence_db/src/entity/channels/channel.controller.ts
@@ -1,11 +1,12 @@
-import { Body, Controller, Get, Param, Post } from '@nestjs/common';
+import { Body, Controller, Get, Param, Post, Req } from '@nestjs/common';
 import { ChannelService } from './channel.service';
 import { ChannelsUsersService } from '../channels_users/channels_users.service';
+import { SessionService } from '../sessions/session.service';
 
 
 @Controller('channels')
 export class ChannelController {
-    constructor (private readonly channelService: ChannelService, private readonly channelsUsersService: ChannelsUsersService) {}
+    constructor (private readonly channelService: ChannelService, private readonly channelsUsersService: ChannelsUsersService, private readonly sessionService: SessionService) {}
     /**
      * get channel list of a user
      * 
@@ -13,8 +14,13 @@ export class ChannelController {
      * @returns result of request
      */
     @Get(':id')
-    async find(@Param('id') user_id : number)
+    async find(@Param('id') user_id : number, @Req() req)
     {
+        if (!req.cookies['SESSION_KEY'] || !this.sessionService.getIsSessionExpired(req.cookies['SESSION_KEY']))
+        {
+            return (null);
+            // TODO redirect to log page
+        }
         const channels = await this.channelsUsersService.findChannelsOfUsers(user_id);
         return (channels);
     }
@@ -25,8 +31,13 @@ export class ChannelController {
      * @returns result of request
      */
     @Post('setPassword')
-    async setPassword(@Body() body)
+    async setPassword(@Body() body, @Req() req)
     {
+        if (!req.cookies['SESSION_KEY'] || !this.sessionService.getIsSessionExpired(req.cookies['SESSION_KEY']))
+        {
+            return (null);
+            // TODO redirect to log page
+        }
         return (await this.channelService.setPassword(body.channel, body.password));
     }
     /**
@@ -36,14 +47,24 @@ export class ChannelController {
      * @returns result of request
      */
     @Post('removePassword')
-    async removePassword(@Body() body)
+    async removePassword(@Body() body, @Req() req)
     {
+        if (!req.cookies['SESSION_KEY'] || !this.sessionService.getIsSessionExpired(req.cookies['SESSION_KEY']))
+        {
+            return (null);
+            // TODO redirect to log page
+        }
         return (await this,this.channelService.removePassword(body.channel));
     }
 
     @Post('changePassword')
-    async changePassword(@Body() body)
+    async changePassword(@Body() body, @Req() req)
     {
+        if (!req.cookies['SESSION_KEY'] || !this.sessionService.getIsSessionExpired(req.cookies['SESSION_KEY']))
+        {
+            return (null);
+            // TODO redirect to log page
+        }
         return (await this,this.channelService.changePassword(body.channel, body.password));
     }
 }
